Improved point-processing with iterators, moving thresholding to use newer style voxel iteration.

+++ anchor-image/src/main/java/org/anchoranalysis/image/extent/ImageDim.java
@@ -89,7 +89,7 @@ public class ImageDim implements Serializable {
 		res.scaleXY(sf);
 	}
 	
-	public int getVolume() {
+	public long getVolume() {
 		return extent.getVolume();
 	}
