+++ anchor-mpp/src/main/java/org/anchoranalysis/anchor/mpp/probmap/ProbMapObjMaskCollection.java
@@ -65,7 +65,7 @@ public class ProbMapObjMaskCollection extends ProbMap {
 		//  on.  Could be very inefficient for low-density bounding boxes? So we should make sure
 		//  bounding boxes are tight
 		
-		int vol = om.getVoxelBox().extnt().getVolume();
+		long vol = om.getVoxelBox().extnt().getVolume();
 		int volXY = om.getVoxelBox().extnt().getVolumeXY();
 		int exY = om.getVoxelBox().extnt().getX();
